tools/10-vibe-coder/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "openclaw-vibe-coder"
version = "1.0.0"
description = "Vibe Coder Dashboard — CodeQwen, DocuWriter, GitAssistant, HKDevKit"
requires-python = ">=3.11"
dependencies = [
    "openclaw-shared",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "python-multipart>=0.0.12",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "psutil>=5.9",
    # CodeQwen
    "sse-starlette>=2.0",
    # DocuWriter
    "tree-sitter>=0.21",
    "mdformat>=0.7",
    # GitAssistant
    "gitpython>=3.1",
    # HKDevKit
    "qrcode[pil]>=7.4",
    "Pillow>=10.0",
]

[project.optional-dependencies]
mlx = ["openclaw-shared[mlx]"]
messaging = ["openclaw-shared[messaging]"]
macos = ["openclaw-shared[macos]"]
all = ["openclaw-vibe-coder[mlx,messaging,macos]"]

[tool.setuptools.packages.find]
include = ["vibe_coder*"]
